requirements.txt
black
python-dotenv
requests
types-requests
